+++ src/oca/go/src/goca/document.go
@@ -37,7 +37,7 @@ type DocumentPool struct {
 
 // Document represents an OpenNebula Document
 type Document struct {
-	ID          uint             `xml:"ID"`
+	ID          int              `xml:"ID"`
 	UID         int              `xml:"UID"`
 	GID         int              `xml:"GID"`
 	UName       string           `xml:"UNAME"`
@@ -59,13 +59,13 @@ func (c *Controller) Documents(dType int) *DocumentsController {
 }
 
 // Document returns a Document controller
-func (c *Controller) Document(id uint) *DocumentController {
+func (c *Controller) Document(id int) *DocumentController {
 	return &DocumentController{c, id}
 }
 
 // ByName returns a Document ID from name
-func (dc *DocumentsController) ByName(name string, args ...int) (uint, error) {
-	var id uint
+func (dc *DocumentsController) ByName(name string, args ...int) (int, error) {
+	var id int
 
 	documentPool, err := dc.Info(args...)
 	if err != nil {
@@ -142,13 +142,13 @@ func (dc *DocumentController) Info() (*Document, error) {
 }
 
 // Create allocates a new document. It returns the new document ID.
-func (dc *DocumentsController) Create(tpl string) (uint, error) {
+func (dc *DocumentsController) Create(tpl string) (int, error) {
 	response, err := dc.c.Client.Call("one.document.allocate", tpl)
 	if err != nil {
 		return 0, err
 	}
 
-	return uint(response.BodyInt()), nil
+	return response.BodyInt(), nil
 }
 
 // Clone clones an existing document.
