+++ src/oca/go/src/goca/template.go
@@ -34,7 +34,7 @@ type TemplatePool struct {
 
 // Template represents an OpenNebula Template
 type Template struct {
-	ID          uint             `xml:"ID"`
+	ID          int              `xml:"ID"`
 	UID         int              `xml:"UID"`
 	GID         int              `xml:"GID"`
 	UName       string           `xml:"UNAME"`
@@ -90,13 +90,13 @@ func (c *Controller) Templates() *TemplatesController {
 }
 
 // Template returns a Template controller
-func (c *Controller) Template(id uint) *TemplateController {
+func (c *Controller) Template(id int) *TemplateController {
 	return &TemplateController{c, id}
 }
 
 // ByName returns a Template by Name
-func (c *TemplatesController) ByName(name string, args ...int) (uint, error) {
-	var id uint
+func (c *TemplatesController) ByName(name string, args ...int) (int, error) {
+	var id int
 
 	templatePool, err := c.Info(args...)
 	if err != nil {
@@ -169,13 +169,13 @@ func (tc *TemplateController) Info() (*Template, error) {
 }
 
 // Create allocates a new template. It returns the new template ID.
-func (tc *TemplatesController) Create(template string) (uint, error) {
+func (tc *TemplatesController) Create(template string) (int, error) {
 	response, err := tc.c.Client.Call("one.template.allocate", template)
 	if err != nil {
 		return 0, err
 	}
 
-	return uint(response.BodyInt()), nil
+	return response.BodyInt(), nil
 }
 
 // Update replaces the cluster cluster contents.
@@ -214,14 +214,14 @@ func (tc *TemplateController) Delete() error {
 }
 
 // Instantiate will instantiate the template
-func (tc *TemplateController) Instantiate(name string, pending bool, extra string, clone bool) (uint, error) {
+func (tc *TemplateController) Instantiate(name string, pending bool, extra string, clone bool) (int, error) {
 	response, err := tc.c.Client.Call("one.template.instantiate", tc.ID, name, pending, extra, clone)
 
 	if err != nil {
 		return 0, err
 	}
 
-	return uint(response.BodyInt()), nil
+	return response.BodyInt(), nil
 }
 
 // Clone an existing template. If recursive is true it will clone the template
